+++ picodi/_state.py
@@ -6,18 +6,17 @@ import threading
 from collections.abc import AsyncGenerator, Awaitable, Callable, Generator
 from contextlib import asynccontextmanager, contextmanager
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, Any, ContextManager, TypeVar, overload
+from typing import TYPE_CHECKING, Any, AsyncContextManager, ContextManager, TypeVar
 
 from picodi._internal import LazyResolver
 from picodi._scopes import AutoScope, ManualScope, NullScope, ScopeType
-from picodi.support import ExitStack, NullAwaitable
+from picodi.support import ExitStack, NullAwaitable, is_async_function
 
 if TYPE_CHECKING:
     from picodi._types import DependencyCallable, InitDependencies, LifespanScopeClass
 
 
 TC = TypeVar("TC", bound=Callable)
-unset = object()
 
 
 class Storage:
@@ -224,26 +223,14 @@ class Registry:
         """
         return frozenset(self._storage.touched_dependencies)
 
-    @overload
-    def override(
-        self, dependency: DependencyCallable, new_dependency: None = None
-    ) -> Callable[[DependencyCallable], DependencyCallable]:
-        pass
-
-    @overload
-    def override(
-        self, dependency: DependencyCallable, new_dependency: DependencyCallable
-    ) -> ContextManager[None]:
-        pass
-
     def override(
         self,
         dependency: DependencyCallable,
-        new_dependency: DependencyCallable | None | object = unset,
-    ) -> Callable[[DependencyCallable], DependencyCallable] | ContextManager[None]:
+        new_dependency: DependencyCallable | None,
+    ) -> ContextManager[None]:
         """
-        Override a dependency with a new one. It can be used as a decorator,
-        as a context manager or as a regular method call. New dependency will be
+        Override a dependency with a new one. It can be used as a context manager
+        or as a regular method call. New dependency will be
         added to the registry.
 
         :param dependency: dependency to override
@@ -254,11 +241,6 @@ class Registry:
         --------
         .. code-block:: python
 
-            @registry.override(get_settings)
-            def real_settings():
-                return {"real": "settings"}
-
-
             with registry.override(get_settings, real_settings):
                 pass
 
@@ -268,20 +250,13 @@ class Registry:
         if self._storage.get_original(dependency):
             raise ValueError("Cannot override an overridden dependency")
 
-        def decorator(override_to: DependencyCallable) -> DependencyCallable:
-            self._storage.add(override_to)
-            if dependency is override_to:
-                raise ValueError("Cannot override a dependency with itself")
-            self._storage.overrides[dependency] = override_to
-            return override_to
-
-        if new_dependency is unset:
-            return decorator
-
         with lock:
             call_dependency = self._storage.overrides.get(dependency)
-            if callable(new_dependency):
-                decorator(new_dependency)
+            if new_dependency is not None:
+                self._storage.add(new_dependency)
+                if dependency is new_dependency:
+                    raise ValueError("Cannot override a dependency with itself")
+                self._storage.overrides[dependency] = new_dependency
             else:
                 self._storage.overrides.pop(dependency, None)
 
@@ -307,7 +282,7 @@ class Registry:
         """
         self._storage.touched_dependencies.clear()
 
-    def clear(self) -> None:
+    def _clear(self) -> None:
         """
         Clear the registry. It will remove all dependencies, overrides
         and touched dependencies.
@@ -319,6 +294,7 @@ class Registry:
         self._storage.deps.clear()
         self._storage.overrides.clear()
         self._storage.touched_dependencies.clear()
+        self._for_init.clear()
 
 
 @dataclass(frozen=True)
@@ -333,12 +309,9 @@ class Provider:
         dependency: DependencyCallable,
         scope: ScopeType,
     ) -> Provider:
-        is_async = inspect.iscoroutinefunction(
-            dependency
-        ) or inspect.isasyncgenfunction(dependency)
         return cls(
             dependency=dependency,
-            is_async=is_async,
+            is_async=is_async_function(dependency),
             scope=scope,
         )
 
@@ -362,6 +335,13 @@ class Provider:
                         assert exit_stack is not None, "exit_stack is required"
                         return await scope.enter(exit_stack, context_manager())
                     return await scope.enter(context_manager())
+                elif isinstance(value_or_gen_, AsyncContextManager):
+                    if isinstance(scope, AutoScope):
+                        assert exit_stack is not None, "exit_stack is required"
+                        return await scope.enter(
+                            exit_stack, _recreate_cm(value_or_gen_)
+                        )
+                    return await scope.enter(_recreate_cm(value_or_gen_))
                 return value_or_gen_
 
             return resolve_value_inner()
@@ -372,7 +352,18 @@ class Provider:
                 assert exit_stack is not None, "exit_stack is required"
                 return scope.enter(exit_stack, context_manager())
             return scope.enter(context_manager())
+        elif isinstance(value_or_gen, ContextManager):
+            if isinstance(scope, AutoScope):
+                assert exit_stack is not None, "exit_stack is required"
+                return scope.enter(exit_stack, _recreate_cm(value_or_gen))
+            return scope.enter(_recreate_cm(value_or_gen))
         return value_or_gen
+
+
+def _recreate_cm(
+    gen: AsyncContextManager | ContextManager,
+) -> AsyncContextManager | ContextManager:
+    return gen._recreate_cm()  # type: ignore[union-attr]  # noqa: SF01
 
 
 lock = threading.RLock()