+++ CHANGELOG.md
@@ -6,12 +6,14 @@ We follow [Semantic Versions](https://semver.org/).
 
 ## Version 0.32.0
 
+- Picodi now can use `contextlib` dependencies like yield functions
 - Changed state management of registry
 - Breaking changes:
   - renamed `helpers.enter` to `helpers.resolve`
   - deleted `dependency`, `init_dependencies`, `shutdown_dependencies` in favor of
     `registry.set_scope`, `registry.init`, `registry.shutdown`
   - deleted `helpers.lifespan` in favor of `registry.lifespan` and `registry.alifespan`
+  - registry.override now can't be used as decorator
 
 ## Version 0.31.0